Refactor into modules.
